tiny changes-
Hey girls - so design-wise I had and idea to have the white text of h1
on the green grass of the background-image somehow -

Hasn’t quite worked yet but thought I would commit the changes so far.
I tried to make a separate div class for it totally didn’t work, if
either of you have any ideas let me know. I’ll be working more on it
tomorrow.